Take the default size for new tmux sessions
Per the man page, the default size for a session would be 80 by 24, but
in practice, your session's size fits your monitor when you attach.

I've found that I get errors when the size is determined for me on a
machine that I'm SSHed into - seems like this bit of config is overkill.